test(examples-chat): extract sendPromptAndWait helper, wait for data-streaming=false
Asserting on intermediate streaming-state DOM is the other source of e2e
flake. The chat composition flips chat-message[data-streaming] to 'false'
when the agent's isLoading() goes false; helper waits on that DOM contract
before returning the finalized bubble. Smoke, markdown, and A2UI specs
all route through the helper now.
